rl-qos-system/requirements.txt
# Python Requirements for RL-QoS System

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
pandas>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Logging
loguru>=0.7.0

# Note: For Linux deployment with Mininet, additional packages needed:
# - ryu (SDN controller)
# - mininet (network emulator)
# - openvswitch (software switch)
# These must be installed via apt/system package manager on Linux
